backend/requirements.txt
Flask
flask-cors
flask-jwt-extended
mysql-connector-python
python-dotenv